requirements.txt
Django
recipe-scrapers
pandas
pyyaml
python-slugify
